More Code Updates
-Added more enums to make the code more readable
-Fixed a possible bug where you were given the drunk cutscene regardless of whether the character was drunk needs testing

+++ Assets/Scripts/Darts/BoardCollider.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BoardCollider : MonoBehaviour
 {
-    public byte point;
+    public int point;
     public DartGame gameState;
     public MeshRenderer mr;
     public Transform target;
